feat: data structure streaming with pyspark
This is a small sample of data streaming, using as input a kafka topic, and then aggregation of data using pyspark, using as sink the console. for the watermark of the input, we are using 30s and a window time of 60min. Also you can find some examples of the expected output in the README.md file. In order to make it work, we need to run our previous kafka_server.